boomshakalaka

+++ SqueletteImplantation/wwwroot/app/pageRechercheEnreprise.component.ts
@@ -86,7 +86,9 @@ export class pageRechercheEntrepriseComponent
         }
     }
    
-    Supprimer(ID:number) {
+    Supprimer(ID:number,Nom:string) {
+        var r = confirm("Voulez-vous supprimer l'entreprise "+Nom.toString());
+    if (r == true) {
         this.http.delete("api/Entreprise/Supprimer/" + ID.toString()).subscribe( donnee => {
                 if (donnee.status == 200) {
                     this.jBoxMessage("green", "Entreprise supprimée avec succès!");
@@ -96,7 +98,7 @@ export class pageRechercheEntrepriseComponent
                     this.jBoxMessage("red", "Erreur lors de la suppression de l'entreprise!");
                 }
             });
-    }
+    }}
 
     AjouterAnnee(entreprise: Entreprise)
     {
